add missing benchmarks from AddDefaultSize, remove AddDefaultSize to avoid duplication
